Add data.rs
Add a file for computing the data costs.  So far it has been filled with a squared difference function, and a function to compute the data costs for the input images.

+++ rust/src/common.rs
@@ -5,7 +5,7 @@ pub struct Parameters {
     pub width: u32,
     pub height: u32,
     pub total_pixels: u32,
-    pub window_edge_size: usize,
-    pub maximum_disparity: usize,
+    pub window_edge_size: u32,
+    pub maximum_disparity: u32,
     pub use_zero_mean: bool,
 }
